+++ src/main/java/de/muspellheim/commons/fx/demo/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Muspellheim Commons FX
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+/**
+ * Contains a demo app for the features og this library.
+ */
+package de.muspellheim.commons.fx.demo;