+++ src/main/java/de/muspellheim/commons/fx/validation/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Muspellheim Commons FX
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+/**
+ * Contains extensions to ControlFX validation.
+ */
+package de.muspellheim.commons.fx.validation;